Make use of serde::ser::Serialize
The essential functionality of `create_table` is implemented by calling
into `serde_json::to_string`. The bounds on this function *require* that
its argument implement `Serialize`, and so `create_table` must demand at
least that. It turns out we need no additional power, to `T: Serialize`
is the correct bound.