Make signatures more idiomatic
These changes give us type signatures that Rustaceans are more likely to
expect.

Since all we need of the `Serializable` parameter to `create_table` is a
reference we can pass to `serde_json::to_string`, it would be rude to
take ownership of it.

Essentially all filesystem-related functions make use of `AsRef<Path>`,
which provides the caller additional flexibility. We might as well
follow their example. This also opens up the use of `Path::join`, which
is good because treating paths as strings gives me hives.