Additional idiomatic changes
Once we use `Path::join` in `create_table`, `db_table` becomes a
`PathBuf` and `exists` is available directly.

Since what we're writing to `buffer` (actually a handle to the
underlying file) is already a string, we don't need to go through the
formatting machinery; we can blit the bytes directly.