Create an error type
This does not currently have fe_bucket specific problems in it; I was
focused on wiring io and serde together.

The unfortunate bit of this design is that there's no distinction
between serialization errors and deserialization errors. That's
inherited from serde_json; we could offer more precise reporting, but
that means giving up try! for serde operations.

ref #10