Add error for nonexistent table and excise panics
The only library-specific error case I noticed in the current code is
trying to operate on a table that isn't there. That error variant
carries the name of the table around as an owned String; we could use a
slice there, but then Error would need a lifetime parameter.

Fallible operations return our Result, instead of possibly panicking.
The most invasive change to do that was read_table, which now reads the
file directly into a string rather than reading line-by-line into an
intermediate buffer and joining at the end, since I didn't want to try
to remember how Rust pronounces "sequence". This version saves us some
copies anyway.

One side-effect of the depanicking is that we don't need
allow(unused_must_use) anymore: all must_use intermediates are returned
to the caller. Except for in tests, which got _more_ panicky, but let's
be honest if any of those steps failed we wanted the test to fail anyway
right?