Exterminate nightly dependency
... at least in theory. In practice this code triggers an internal
compiler error on stable (1.9.0), which makes it not so useful as I was
hoping.

The serialization and deserialization code is quite long and entirely
mechanical, as might be expected from a direct port of semantics we used
to get from a compiler plugin. On the other hand we can now start
thinking about the on-disk format separately from the in-memory format,
which is a win in my book.